backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pgvector==0.3.2
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
django-environ==0.11.2
Pillow==10.2.0
openai>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
numpy>=1.24.0
